x86: intel_sgx: add batch allocation of EPC pages
Add a function to allocate a batch of EPC pages in a single shot.
KVM will use this to pre-allocate a large number of EPC pages on
behalf of the guest prior to launching the guest.  This allows KVM
to balance robustness and performance, e.g. early detection of EPC
OOM versus lazily installing EPC pages.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>